#764 Service for downloading submissions from the browser
